REF: deprecate mv_interface in favor of interface.
The features in mv_interface have expanded past the original scope.
The new name better reflects the content of the module while maintaining
a grouping of similar features.
