+++ pcdsdevices/lens.py
@@ -8,7 +8,7 @@ from ophyd.pseudopos import (PseudoPositioner, PseudoSingle,
 from .doc_stubs import basic_positioner_init
 from .epics_motor import IMS
 from .inout import InOutRecordPositioner
-from .mv_interface import tweak_base
+from .interface import tweak_base
 from .sim import FastMotor
