+++ pcdsdevices/state.py
@@ -12,7 +12,7 @@ from ophyd.device import Device, Component as Cpt
 
 from .doc_stubs import basic_positioner_init
 from .epics_motor import IMS
-from .mv_interface import MvInterface
+from .interface import MvInterface
 from .signal import AggregateSignal
 
 logger = logging.getLogger(__name__)
